fix drop down bug
